Restructure under src directory
