+++ pyproject.toml
@@ -5,8 +5,11 @@ build-backend = "hatchling.build"
 [project]
 name = "transform_utils"
 version = "0.0.1"
-dependencies = []
+dependencies = ["numpy"]
 authors = [{ name = "Benned Hedegaard", email = "bhedegaa@cs.brown.edu" }]
 description = "Utility classes for representing and converting spatial transforms."
 readme = "README.md"
 license = { file = "LICENSE" }
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/transform_utils"] # Include packages at these paths
