+++ src/transform_utils/world_model/object_model.py
